requirements.txt
matrix-nio
pyyaml